Look up country for postcode
This works by generating a trie of postcode digits to country, then
truncating it wherever any path can only lead to the same country.

The truncated trie is then filtered by country and turned into a regular
expression. Regular expressions are fast and much denser than a
hash-based trie (about 1/12 as many bytes, avoiding a massive gem.

A Makefile is included to automate fetching and building the data
needed.