Add Post Town Lookup Functionality
Actions:

+ Added two sets of JSON data that store the post
town and the regex to match it.

+ Added a town_finder class, with logic to lookup
the post-town from the JSON data objects.  It
will return either the name of the post town, as
a string, :invalid_postcode (based on existing
logic), :unknown_outcode (in the event that the
postcode appears valid, but no town has the
outcode), or :multiple_possible_matches (if the
postcode is incomplete, and could belong to one or
more post-towns).

+ Added a town method to abstract_postcode and
geographic_postcode.

+ Created a town_finder_spec, and updated the
full_postcode_spec to work with a csv file that
has a town column.

+ Updated country_lookup to separate out Jersey
and Guernsey, based on 'JE' or 'GY' postcode area.
Amended tests in full_postcode_spec to test this.

+ Bumped gem version

Motivations and notes:

+ This provides additional, valuable functionality
to the gem.

+ It will provide the post-town of a postcode,
based on the postcode area (in most instances this
will be based on the outward postcode, but in
some cases, these areas are shared, so the matcher
will check into the inward postcode).

+ These areas were checked against Wikipedia
postcode area lists
https://en.wikipedia.org/wiki/List_of_postcode_areas_in_the_United_Kingdom
and were verified against other publically
availale sources.

+ It does not guarantee the validity of the
postcode.

+ The regexes will need to be updated when new
postcode areas are added, but these changes are
published.